refactor(cms): split editor defaults
Move static CMS editor defaults and validation limits into a dedicated defaults module so cms/index.ts contains normalization and query behavior instead of embedded fallback inventory.

Verified with pnpm --filter @anipotts/lib typecheck; pnpm --filter @anipotts/lib test; pnpm dlx knip --no-exit-code; pnpm exec prettier --check packages/lib/src/cms/defaults.ts packages/lib/src/cms/index.ts; git diff --check.
